requirements.txt
Django==3.0.3
django-allauth==0.38.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
django-rest-auth==0.9.3
djangorestframework==3.11.0
gunicorn==19.10.0
oauthlib==2.0.3
psycopg2==2.8.4
requests==2.22.0
requests-oauthlib==0.8.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
Pillow==7.0.0
beautifulsoup4==4.9.0
django-rest-swagger==2.2.0
